+++ biz/content/src/main/java/com/mike/novel/content/service/TasksService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.mike.novel.dto.TasksDo;
 
 public interface TasksService {
-	List<TasksDo> queryByNid(long nid);
+	List<TasksDo> queryByNid(int nid);
 
 	Integer save(TasksDo tasksDo);
